+++ packages/fetch/CHANGELOG.md
@@ -1,5 +1,11 @@
 # dredge-fetch
 
+## 0.5.3
+
+### Patch Changes
+
+- Fix dependecies in package.json
+
 ## 0.5.2
 
 ### Patch Changes